initial migration from cypress to jest-puppeteer
