+++ e2e-tests/production-runtime/.gitignore
@@ -6,7 +6,3 @@ yarn-error.log
 # Build directory
 /public
 .DS_Store
-
-# Cypress output
-cypress/videos/
-cypress/screenshots/
