+++ e2e-tests/production-runtime/__tests__/1-production.js
@@ -1,0 +1,8 @@
+/* globals page */
+
+describe(`Production build tests`, () => {
+  it(`should render properly`, async () => {
+    await page.goto(`https://www.google.com/`)
+    expect(await page.title()).toMatch(`Google`)
+  })
+})
