--- e2e-tests/production-runtime/cypress.json
@@ -1,4 +1,0 @@
-{
-  "baseUrl": "http://localhost:9000",
-  "chromeWebSecurity": false
-}
