+++ e2e-tests/production-runtime/jest-puppeteer.config.js
@@ -1,0 +1,9 @@
+module.exports = {
+  server: {
+    command: `npm run serve`,
+    port: 9000,
+  },
+  launch: {
+    headless: false,
+  },
+}
