+++ e2e-tests/production-runtime/jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  preset: `jest-puppeteer`,
+  testPathIgnorePatterns: [`node_modules`, `.cache`],
+}
